Check and wait for job completion
At the moment GitLab CI quits as soon as all the test jobs have been
submitted. This means we are dependant on checking LAVA directly to see if
there were any test errors or not.

This commit makes changes so that GitLab CI will wait until every job is
completed.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>