Improve localversion pickup
The get_kernel_name() function creates a name based on the Kernel version
(make kernelversion), contents of the "localversion-cip" file plus the
commit SHA.

The problem with the current approach is that it breaks if there are more
than one "localversion*" file, for example a "localversion-cip" and
"localversion-cip-rt".

This new method can cope with multiple local version files.
Example output: 4.19.50-cip3-rt1_b77ccf6e9d06

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>