Add support for building for x86
This patch includes a lot of changes as the previous approach assumed that
all devices had a device tree, which x86 does not.

All changes should still be backwards compatible though, so no need to
increase the DOCKER_IMAGE_TAG version.

Currently it is not possible to test an x86 board as a healthcheck
template hasn't been added yet. This will be done once one has been added
to cip-lava-docker.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>