Select Kernel config based on direct path
At the moment in the build_kernel.sh script we search for a specified
Kernel configuration in the cip-kernel-config repository using find.

This can cause an issue if there is more than one configuration with the
same filename. For example, there is a toshiba_defconfig for both the x86
and powerpc architectures.

As we already know the full path to get to the relevant config
(version/arch/), we may as well check if the file is present directly.

Fixes #1.

Reported-by: Zoran Stojsavljevic <zoran.stojsavljevic.de@gmail.com>
Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>