Exit immediately on submission failure.
Due to the various helper functions and variable locality it is a little
hard to follow the error handling flow. In any case, I found that if the
submission fails on upload (URL error or login credential failure, etc)
then the script mentions the failure but does not later on exit with an
error code.  This means the CI system wrongly reports success.

A quick solution is to exit immediately with error, as done here.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>